Refactor folders names
