+++ src/View/Album/gallery.html.twig
@@ -20,3 +20,11 @@
     </div>
 </section>
 {% endblock %}
+
+{% block stylesheet %}
+<link rel="stylesheet" href="/assets/plugins/featherlight-1.7.13/featherlight.min.css"/>
+{% endblock %}
+
+{% block javascript %}
+<script type="text/javascript" src="/assets/plugins/featherlight-1.7.13/featherlight.min.js"></script>
+{% endblock %}
